Refactor CovidCollectionDocument to be abstract base class
